UI 수정, 공동 상,하단 템플릿 적용

+++ src/main/webapp/jsp/usr/article/list.jsp
@@ -5,18 +5,20 @@
 <%
     List<ArticleDto> articles = (List<ArticleDto>)request.getAttribute("articleDtos");
 %>
-
-<h1>게시물 리스트</h1>
-
-<ul>
-    <% for ( ArticleDto article : articles ) { %>
-        <li>
-            <a href="/usr/article/detail/free/<%=article.getId()%>"><%=article.getId()%>. <%=article.getTitle()%></a>
-            <br>
-
-            <a href="/usr/article/delete/free/<%=article.getId()%>">삭제</a>
-            <a href="/usr/article/modify/free/<%=article.getId()%>">수정</a>
-        </li>
-
-    <% } %>
-</ul>
+<%@ include file="../common/head.jspf"%>
+<section>
+    <div class="container px-3 mx-auto">
+        <h1 class="font-bold text-lg">게시물 리스트</h1>
+        <ul class="mt-5">
+            <% for ( ArticleDto article : articles ) { %>
+            <li class="flex">
+                <a class="w-[40px] hover:underline hover:text-[red]" href="/usr/article/detail/free/<%=article.getId()%>"><%=article.getId()%></a>
+                <a class="flex-grow hover:underline hover:text-[red]" href="/usr/article/detail/free/<%=article.getId()%>"><%=article.getTitle()%></a>
+                <a onclick="if ( !confirm('정말로 삭제하시겠습니까?') ) return false;" class="hover:underline hover:text-[red] mr-2" href="/usr/article/delete/free/<%=article.getId()%>">삭제</a>
+                <a class="hover:underline hover:text-[red]" href="/usr/article/modify/free/<%=article.getId()%>">수정</a>
+            </li>
+            <% } %>
+        </ul>
+    </div>
+</section>
+<%@ include file="../common/foot.jspf"%>
